Moved the convert classes we want to keep to the core folder. The rest will be deleted.
