--- nw/convert/__init__.py
@@ -1,27 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from nw.convert.tokenizer import Tokenizer
-
-from nw.convert.file.concat import ConcatFile
-from nw.convert.file.html import HtmlFile
-from nw.convert.file.latex import LaTeXFile
-from nw.convert.file.markdown import MarkdownFile
-from nw.convert.file.text import TextFile
-
-from nw.convert.text.tohtml import ToHtml
-from nw.convert.text.tolatex import ToLaTeX
-from nw.convert.text.tomarkdown import ToMarkdown
-from nw.convert.text.totext import ToText
-
-__all__ = [
-    "Tokenizer",
-    "ConcatFile",
-    "HtmlFile",
-    "LaTeXFile",
-    "MarkdownFile",
-    "TextFile",
-    "ToHtml",
-    "ToLaTeX",
-    "ToMarkdown",
-    "ToText",
-]
