+++ nw/core/__init__.py
@@ -6,6 +6,8 @@ from nw.core.project import NWProject
 from nw.core.spellcheck import NWSpellCheck
 from nw.core.spellcheck import NWSpellEnchant
 from nw.core.spellcheck import NWSpellSimple
+from nw.core.tokenizer import Tokenizer
+from nw.core.tohtml import ToHtml
 from nw.core.tools import countWords
 from nw.core.tools import projectMaintenance
 from nw.core.tools import numberToWord
@@ -17,6 +19,8 @@ __all__ = [
     "NWSpellCheck",
     "NWSpellEnchant",
     "NWSpellSimple",
+    "Tokenizer",
+    "ToHtml",
     "countWords",
     "projectMaintenance",
     "numberToWord",
