+++ nw/gui/dialogs/__init__.py
@@ -3,7 +3,6 @@
 from nw.gui.dialogs.configeditor import GuiConfigEditor
 from nw.gui.dialogs.docmerge import GuiDocMerge
 from nw.gui.dialogs.docsplit import GuiDocSplit
-from nw.gui.dialogs.export import GuiExport
 from nw.gui.dialogs.itemeditor import GuiItemEditor
 from nw.gui.dialogs.projecteditor import GuiProjectEditor
 from nw.gui.dialogs.projectload import GuiProjectLoad
@@ -13,7 +12,6 @@ __all__ = [
     "GuiConfigEditor",
     "GuiDocMerge",
     "GuiDocSplit",
-    "GuiExport",
     "GuiItemEditor",
     "GuiProjectEditor",
     "GuiProjectLoad",
